requirements.txt
Django==3.2.8  
djangorestframework==3.12.4 
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0 
psycopg2==2.9.1
django-filter==21.1
Pyrebase4==4.5.0
gunicorn==20.1.0
django-heroku==0.3.
django-cors-headers==3.9.0